+++ src/components/CustomModal.vue
@@ -29,7 +29,12 @@
         <hr />
         <div class="w-full flex justify-center md:justify-end pt-2 md:pt-0">
           <button
-            @click="$emit('close')"
+            @click="
+              () => {
+                $emit('close');
+                $emit('confirm');
+              }
+            "
             class="
               bg-transparent
               hover:bg-yellow-300
